+++ src/main/kotlin/org/acejump/view/TextHighlighter.kt
@@ -1,28 +1,21 @@
 package org.acejump.view
 
 import com.intellij.codeInsight.CodeInsightBundle
-import com.intellij.codeInsight.hint.HintManagerImpl
-import com.intellij.codeInsight.hint.HintUtil
+import com.intellij.codeInsight.hint.*
+import com.intellij.codeInsight.hint.HintManagerImpl.HIDE_BY_ESCAPE
+import com.intellij.codeInsight.hint.HintManagerImpl.HIDE_BY_TEXT_CHANGE
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.editor.colors.EditorFontType
-import com.intellij.openapi.editor.markup.CustomHighlighterRenderer
-import com.intellij.openapi.editor.markup.HighlighterLayer
-import com.intellij.openapi.editor.markup.HighlighterTargetArea
-import com.intellij.openapi.editor.markup.RangeHighlighter
-import com.intellij.ui.HintHint
-import com.intellij.ui.JBColor
-import com.intellij.ui.LightweightHint
-import com.intellij.util.ui.JBUI
-import com.intellij.util.ui.UIUtil
+import com.intellij.openapi.editor.markup.*
+import com.intellij.openapi.editor.markup.HighlighterTargetArea.EXACT_RANGE
+import com.intellij.ui.*
+import com.intellij.util.ui.*
 import it.unimi.dsi.fastutil.ints.IntList
+import org.acejump.*
 import org.acejump.boundaries.EditorOffsetCache
 import org.acejump.config.AceConfig
-import org.acejump.immutableText
 import org.acejump.input.JumpMode
-import org.acejump.isWordPart
 import org.acejump.search.SearchQuery
-import org.acejump.wordEnd
-import org.acejump.wordStart
 import java.awt.*
 import javax.swing.*
 import kotlin.math.max
@@ -31,26 +24,24 @@ import kotlin.math.max
  * Renders highlights for search occurrences.
  */
 internal class TextHighlighter {
-  private companion object {
-    private const val LAYER = HighlighterLayer.LAST + 1
-  }
-
+  private companion object { private const val LAYER = HighlighterLayer.LAST + 1 }
   private var previousHighlights = mutableMapOf<Editor, Array<RangeHighlighter>>()
-  private var previousHint : LightweightHint? = null
+  private var previousHint: LightweightHint? = null
 
   /**
    * Label for the search notification.
    */
-  private class NotificationLabel constructor(text: String?) : JLabel(text) {
+  private class NotificationLabel constructor(text: String?): JLabel(text) {
     init {
       background = HintUtil.getInformationColor()
       foreground = JBColor.foreground()
       this.isOpaque = true
     }
   }
-  
+
   /**
-   * Removes all current highlights and re-creates them from scratch. Must be called whenever any of the method parameters change.
+   * Removes all current highlights and re-creates them from scratch.
+   * Must be called whenever any of the method parameters change.
    */
   fun render(results: Map<Editor, IntList>, query: SearchQuery, jumpMode: JumpMode) {
 
@@ -59,51 +50,48 @@ internal class TextHighlighter {
       jumpMode === JumpMode.TARGET -> SearchedWordWithOutlineRenderer
       else -> SearchedWordRenderer
     }
-    
+
     for ((editor, offsets) in results) {
       val highlights = previousHighlights[editor]
-    
+
       val markup = editor.markupModel
       val document = editor.document
       val chars = editor.immutableText
-    
+
       val modifications = (highlights?.size ?: 0) + offsets.size
       val enableBulkEditing = modifications > 1000
-    
+
       try {
-        if (enableBulkEditing) {
-          document.isInBulkUpdate = true
-        }
-    
+        if (enableBulkEditing) document.isInBulkUpdate = true
+
         highlights?.forEach(markup::removeHighlighter)
         previousHighlights[editor] = Array(offsets.size) { index ->
           val start = offsets.getInt(index)
           val end = start + query.getHighlightLength(chars, start)
-      
-          markup.addRangeHighlighter(start, end, LAYER, null, HighlighterTargetArea.EXACT_RANGE).apply {
-            customRenderer = renderer
-          }
+
+          markup.addRangeHighlighter(start, end, LAYER, null, EXACT_RANGE)
+            .apply { customRenderer = renderer }
         }
       } finally {
         if (enableBulkEditing) document.isInBulkUpdate = false
       }
     }
 
-    if (AceConfig.showSearchNotification) {
+    if (AceConfig.showSearchNotification)
       showSearchNotification(results, query, jumpMode)
-    }
 
     for (editor in previousHighlights.keys.toList()) {
-      if (!results.containsKey(editor)) {
-        previousHighlights.remove(editor)?.forEach(editor.markupModel::removeHighlighter)
-      }
+      if (!results.containsKey(editor))
+        previousHighlights.remove(editor)
+          ?.forEach(editor.markupModel::removeHighlighter)
     }
   }
 
   /**
    * Show a notification with the current search text.
    */
-  private fun showSearchNotification(results: Map<Editor, IntList>, query: SearchQuery, jumpMode: JumpMode) {
+  private fun showSearchNotification(results: Map<Editor, IntList>,
+                                     query: SearchQuery, jumpMode: JumpMode) {
     // clear previous hint
     previousHint?.hide()
 
@@ -111,16 +99,19 @@ internal class TextHighlighter {
     results.keys.first().let {
       val component: JComponent = it.component
 
-      val label1: JLabel = NotificationLabel(" " + CodeInsightBundle.message("incremental.search.tooltip.prefix"))
+      val label1: JLabel = NotificationLabel(" " +
+        CodeInsightBundle.message("incremental.search.tooltip.prefix"))
       label1.font = UIUtil.getLabelFont().deriveFont(Font.BOLD)
 
       val isRegex = query is SearchQuery.RegularExpression
-      val queryText =
-        if (isRegex) " ${query.rawText}" else query.rawText[0] + query.rawText.drop(1).toLowerCase()
-      val label2: JLabel = NotificationLabel(queryText)
+      val queryText = " " + if (isRegex) query.rawText
+      else query.rawText[0] + query.rawText.drop(1).toLowerCase()
+      val label2 = NotificationLabel(queryText)
 
-      val label3: JLabel =
-        NotificationLabel("Found ${results.values.flatMap { it.asIterable() }.size} results in ${results.keys.size} editors.")
+      val label3 = NotificationLabel(
+        "Found ${results.values.flatMap { it.asIterable() }.size}" +
+        "results in ${results.keys.size} editors."
+      )
 
       val panel = JPanel(BorderLayout())
       panel.add(label1, BorderLayout.WEST)
@@ -129,22 +120,21 @@ internal class TextHighlighter {
       panel.border = BorderFactory.createLineBorder(jumpMode.caretColor)
 
       panel.background = jumpMode.caretColor
-      panel.preferredSize = Dimension(
-        it.contentComponent.width + label1.preferredSize.width,
-        panel.preferredSize.height
-      )
+      panel.preferredSize = Dimension(it.contentComponent.width +
+          label1.preferredSize.width, panel.preferredSize.height)
 
       val hint = LightweightHint(panel)
 
       val x = SwingUtilities.convertPoint(component, 0, 0, component).x
       val y: Int = -hint.component.preferredSize.height
-      val p = SwingUtilities.convertPoint(component, x, y, component.rootPane.layeredPane)
+      val p = SwingUtilities.convertPoint(component, x, y,
+        component.rootPane.layeredPane)
 
       HintManagerImpl.getInstanceImpl().showEditorHint(
         hint,
         it,
         p,
-        HintManagerImpl.HIDE_BY_ESCAPE or HintManagerImpl.HIDE_BY_TEXT_CHANGE,
+        HIDE_BY_ESCAPE or HIDE_BY_TEXT_CHANGE,
         0,
         false,
         HintHint(it, p).setAwtTooltip(false)
@@ -179,8 +169,9 @@ internal class TextHighlighter {
   }
 
   /**
-   * Renders a filled highlight in the background of a searched text occurrence, as well as an outline indicating the range of characters
-   * that will be selected by [JumpMode.TARGET].
+   * Renders a filled highlight in the background of a searched
+   * text occurrence, as well as an outline indicating the range
+   * of characters that will be selected by [JumpMode.TARGET].
    */
   private object SearchedWordWithOutlineRenderer: CustomHighlighterRenderer {
     override fun paint(editor: Editor, highlighter: RangeHighlighter, g: Graphics) {
@@ -190,10 +181,7 @@ internal class TextHighlighter {
       val startOffset = highlighter.startOffset
 
       if (chars.getOrNull(startOffset)?.isWordPart == true)
-        drawOutline(
-          g, editor, chars.wordStart(startOffset),
-          chars.wordEnd(startOffset) + 1
-        )
+        drawOutline(g, editor, chars.wordStart(startOffset), chars.wordEnd(startOffset) + 1)
     }
 
     private fun drawOutline(g: Graphics, editor: Editor, startOffset: Int, endOffset: Int) {
@@ -201,12 +189,14 @@ internal class TextHighlighter {
       val end = EditorOffsetCache.Uncached.offsetToXY(editor, endOffset)
 
       g.color = AceConfig.targetModeColor
-      g.drawRect(max(0, start.x - JBUI.scale(1)), start.y, end.x - start.x + JBUI.scale(2), editor.lineHeight)
+      g.drawRect(max(0, start.x - JBUI.scale(1)), start.y,
+        end.x - start.x + JBUI.scale(2), editor.lineHeight)
     }
   }
 
   /**
-   * Renders a filled highlight in the background of the first highlighted position. Used for regex search queries.
+   * Renders a filled highlight in the background of the first highlighted
+   * position. Used for regex search queries.
    */
   private object RegexRenderer: CustomHighlighterRenderer {
     override fun paint(editor: Editor, highlighter: RangeHighlighter, g: Graphics) =